+++ tests/VCardExceptionTest.php
@@ -12,19 +12,16 @@ require_once __DIR__ . '/../vendor/autoload.php';
  * file that was distributed with this source code.
  */
 
-
 /**
  * VCard Exception Test.
- *
- * @author Jeroen Desloovere <info@jeroendesloovere.be>
  */
-class ExceptionTest extends \PHPUnit_Framework_TestCase
+class VCardExceptionTest extends \PHPUnit_Framework_TestCase
 {
     /**
-     * @expectedException Exception
+     * @expectedException JeroenDesloovere\VCard\VCardException
      */
     public function testException()
     {
-        throw new Exception('Testing the VCard error.');
+        throw new VCardException('Testing the VCard error.');
     }
 }