+++ tests/VCardTest.php
@@ -16,8 +16,6 @@ use JeroenDesloovere\VCard\VCard;
 
 /**
  * This class will test our VCard PHP Class which can generate VCards.
- *
- * @author Jeroen Desloovere <info@jeroendesloovere.be>
  */
 class VCardTest extends \PHPUnit_Framework_TestCase
 {
@@ -33,12 +31,12 @@ class VCardTest extends \PHPUnit_Framework_TestCase
      */
     public function emailDataProvider()
     {
-        return array(
-            array(array('john@doe.com')),
-            array(array('john@doe.com', 'WORK' => 'john@work.com')),
-            array(array('WORK' => 'john@work.com', 'HOME' => 'john@home.com')),
-            array(array('PREF;WORK' => 'john@work.com', 'HOME' => 'john@home.com')),
-        );
+        return [
+            [['john@doe.com']],
+            [['john@doe.com', 'WORK' => 'john@work.com']],
+            [['WORK' => 'john@work.com', 'HOME' => 'john@home.com']],
+            [['PREF;WORK' => 'john@work.com', 'HOME' => 'john@home.com']],
+        ];
     }
 
     /**
@@ -48,6 +46,9 @@ class VCardTest extends \PHPUnit_Framework_TestCase
      */
     public function setUp()
     {
+        // set timezone
+        date_default_timezone_set('Europe/Brussels');
+
         $this->vcard = new VCard();
 
         $this->firstName = 'Jeroen';
@@ -89,6 +90,11 @@ class VCardTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($this->vcard, $this->vcard->addCompany(''));
     }
 
+    public function testAddCategories()
+    {
+        $this->assertEquals($this->vcard, $this->vcard->addCategories([]));
+    }
+
     public function testAddEmail()
     {
         $this->assertEquals($this->vcard, $this->vcard->addEmail($this->emailAddress1));
@@ -99,6 +105,11 @@ class VCardTest extends \PHPUnit_Framework_TestCase
     public function testAddJobTitle()
     {
         $this->assertEquals($this->vcard, $this->vcard->addJobtitle(''));
+    }
+
+    public function testAddRole()
+    {
+        $this->assertEquals($this->vcard, $this->vcard->addRole(''));
     }
 
     public function testAddName()
@@ -125,6 +136,80 @@ class VCardTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($this->vcard, $return);
     }
 
+    public function testAddPhotoWithRemoteJpgPhoto()
+    {
+        $return = $this->vcard->addPhoto(
+            'https://raw.githubusercontent.com/jeroendesloovere/vcard/master/tests/image.jpg',
+            true
+        );
+
+        $this->assertEquals($this->vcard, $return);
+    }
+
+    /**
+     * Test adding remote empty photo
+     *
+     * @expectedException Exception
+     * @expectedExceptionMessage Returned data is not an image.
+     */
+    public function testAddPhotoWithRemoteEmptyJpgPhoto()
+    {
+        $this->vcard->addPhoto(
+            'https://raw.githubusercontent.com/jeroendesloovere/vcard/master/tests/empty.jpg',
+            true
+        );
+    }
+
+    public function testAddPhotoContentWithJpgPhoto()
+    {
+        $return = $this->vcard->addPhotoContent(file_get_contents(__DIR__ . '/image.jpg'));
+
+        $this->assertEquals($this->vcard, $return);
+    }
+
+    /**
+     * Test adding empty photo
+     *
+     * @expectedException Exception
+     * @expectedExceptionMessage Returned data is not an image.
+     */
+    public function testAddPhotoContentWithEmptyContent()
+    {
+        $this->vcard->addPhotoContent('');
+    }
+
+    public function testAddLogoWithJpgImage()
+    {
+        $return = $this->vcard->addLogo(__DIR__ . '/image.jpg', true);
+
+        $this->assertEquals($this->vcard, $return);
+    }
+
+    public function testAddLogoWithJpgImageNoInclude()
+    {
+        $return = $this->vcard->addLogo(__DIR__ . '/image.jpg', false);
+
+        $this->assertEquals($this->vcard, $return);
+    }
+
+    public function testAddLogoContentWithJpgImage()
+    {
+        $return = $this->vcard->addLogoContent(file_get_contents(__DIR__ . '/image.jpg'));
+
+        $this->assertEquals($this->vcard, $return);
+    }
+
+    /**
+     * Test adding empty photo
+     *
+     * @expectedException Exception
+     * @expectedExceptionMessage Returned data is not an image.
+     */
+    public function testAddLogoContentWithEmptyContent()
+    {
+        $this->vcard->addLogoContent('');
+    }
+
     public function testAddUrl()
     {
         $this->assertEquals($this->vcard, $this->vcard->addUrl('1'));
@@ -133,21 +218,32 @@ class VCardTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * Test adding photo with no value
+     * Test adding local photo using an empty file
      *
-     * @expectedException JeroenDesloovere\VCard\VCardMediaException
-     * @t@github.com:jeroendesloovere/vcard.gitexpectedExceptionMessage Nothing returned from URL.
+     * @expectedException Exception
+     * @expectedExceptionMessage Returned data is not an image.
      */
-    public function testAddPhotoWithNoValue()
+    public function testAddPhotoWithEmptyFile()
     {
         $this->vcard->addPhoto(__DIR__ . '/emptyfile', true);
     }
 
     /**
+     * Test adding logo with no value
+     *
+     * @expectedException Exception
+     * @expectedExceptionMessage Returned data is not an image.
+     */
+    public function testAddLogoWithNoValue()
+    {
+        $this->vcard->addLogo(__DIR__ . '/emptyfile', true);
+    }
+
+    /**
      * Test adding photo with no photo
      *
-     * @expectedException JeroenDesloovere\VCard\VCardMediaException
-     * @expectedExceptionMessage Returned data aren't an image.
+     * @expectedException Exception
+     * @expectedExceptionMessage Returned data is not an image.
      */
     public function testAddPhotoWithNoPhoto()
     {
@@ -155,11 +251,32 @@ class VCardTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
+     * Test adding logo with no image
+     *
+     * @expectedException Exception
+     * @expectedExceptionMessage Returned data is not an image.
+     */
+    public function testAddLogoWithNoImage()
+    {
+        $this->vcard->addLogo(__DIR__ . '/wrongfile', true);
+    }
+
+    /**
+     * Test charset
+     */
+    public function testCharset()
+    {
+        $charset = 'ISO-8859-1';
+        $this->vcard->setCharset($charset);
+        $this->assertEquals($charset, $this->vcard->getCharset());
+    }
+
+    /**
      * Test Email
      *
      * @dataProvider emailDataProvider $emails
      */
-    public function testEmail($emails = array())
+    public function testEmail($emails = [])
     {
         foreach ($emails as $key => $email) {
             if (is_string($key)) {
@@ -215,7 +332,7 @@ class VCardTest extends \PHPUnit_Framework_TestCase
     /**
      * Test multiple birthdays
      *
-     * @expectedException JeroenDesloovere\VCard\Exception
+     * @expectedException Exception
      */
     public function testMultipleBirthdays()
     {
@@ -224,9 +341,20 @@ class VCardTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
+     * Test multiple categories
+     *
+     * @expectedException Exception
+     */
+    public function testMultipleCategories()
+    {
+        $this->assertEquals($this->vcard, $this->vcard->addCategories(['1']));
+        $this->assertEquals($this->vcard, $this->vcard->addCategories(['2']));
+    }
+
+    /**
      * Test multiple companies
      *
-     * @expectedException JeroenDesloovere\VCard\Exception
+     * @expectedException Exception
      */
     public function testMultipleCompanies()
     {
@@ -237,7 +365,7 @@ class VCardTest extends \PHPUnit_Framework_TestCase
     /**
      * Test multiple job titles
      *
-     * @expectedException JeroenDesloovere\VCard\Exception
+     * @expectedException Exception
      */
     public function testMultipleJobtitles()
     {
@@ -246,9 +374,20 @@ class VCardTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
+     * Test multiple roles
+     *
+     * @expectedException Exception
+     */
+    public function testMultipleRoles()
+    {
+        $this->assertEquals($this->vcard, $this->vcard->addRole('1'));
+        $this->assertEquals($this->vcard, $this->vcard->addRole('2'));
+    }
+
+    /**
      * Test multiple names
      *
-     * @expectedException JeroenDesloovere\VCard\Exception
+     * @expectedException Exception
      */
     public function testMultipleNames()
     {
@@ -259,7 +398,7 @@ class VCardTest extends \PHPUnit_Framework_TestCase
     /**
      * Test multiple notes
      *
-     * @expectedException JeroenDesloovere\VCard\Exception
+     * @expectedException Exception
      */
     public function testMultipleNotes()
     {
@@ -291,5 +430,36 @@ class VCardTest extends \PHPUnit_Framework_TestCase
         );
 
         $this->assertEquals('garcon-jeroen', $this->vcard->getFilename());
+    }
+
+    /**
+     * Test multiple labels
+     */
+    public function testMultipleLabels()
+    {
+        $this->assertSame($this->vcard, $this->vcard->addLabel('My label'));
+        $this->assertSame($this->vcard, $this->vcard->addLabel('My work label', 'WORK'));
+        $this->assertSame(2, count($this->vcard->getProperties()));
+        $this->assertContains('LABEL:My label', $this->vcard->getOutput());
+        $this->assertContains('LABEL;WORK:My work label', $this->vcard->getOutput());
+    }
+
+    public function testChunkSplitUnicode()
+    {
+        $class_handler  = new \ReflectionClass('JeroenDesloovere\VCard\VCard');
+        $method_handler = $class_handler->getMethod('chunk_split_unicode');
+        $method_handler->setAccessible(true);
+
+        $ascii_input="Lorem ipsum dolor sit amet,";
+        $ascii_output = $method_handler->invokeArgs(new VCard(), [$ascii_input,10,'|']);
+        $unicode_input='Τη γλώσσα μου έδωσαν ελληνική το σπίτι φτωχικό στις αμμουδιές του Ομήρου.';
+        $unicode_output = $method_handler->invokeArgs(new VCard(), [$unicode_input,10,'|']);
+
+        $this->assertEquals(
+            "Lorem ipsu|m dolor si|t amet,|",
+            $ascii_output);
+        $this->assertEquals(
+            "Τη γλώσσα |μου έδωσαν| ελληνική |το σπίτι φ|τωχικό στι|ς αμμουδιέ|ς του Ομήρ|ου.|",
+            $unicode_output);
     }
 }